Put UUID into a temporary var when calling uuid_create() and memcpy it into buf_
On HP-UX, when using libdcekt's uuid_create() function, I get bus
errors if we attempt to use buf_ in uuid_create() directly.  It
appears that the buffer needs to be aligned when it is passed through
uuid_create().  So now we just create a temporary uuid_t, pass that to
uuid_create(), and memcpy the result into the buf_, making sure to not
overrun either buffer in the process.

Signed-off-by: AJ Lewis <aj.lewis@quantum.com>